requirements.txt
numpy~=1.23.2
keras~=2.9.0
tensorflow~=2.9.1
nltk~=3.7
PyYAML~=6.0
